ai_matching_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
torch>=2.1.0
torchvision>=0.16.0
ultralytics==8.1.0
transformers==4.37.0
faiss-cpu==1.7.4
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
Pillow==10.2.0
requests==2.31.0
numpy==1.26.3
huggingface-hub==0.20.0
